Refactor: Gatherings 컴포넌트를 클라이언트 컴포넌트로 분리 - page.tsx를 서버 컴포넌트로 유지하도록 클라이언트 컴포넌트 분리
